refactor: extract methods
